// launchdarkly/types/paths/api/v2/flags/{projectKey}/{featureFlagKey}.types.ts
export type getFeatureFlag = (
  $: OmitValueWhenNever<{
    query: getFeatureFlag_Query;
    querystring: never;
    path: getFeatureFlag_Path;
    headers: getFeatureFlag_Headers;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: FeatureFlag;
          };
        };
        examples: {};
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UnauthorizedErrorRep;
          };
        };
        examples: {};
      };
      403: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: ForbiddenErrorRep;
          };
        };
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: NotFoundErrorRep;
          };
        };
        examples: {};
      };
      429: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: RateLimitedErrorRep;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type patchFeatureFlag = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: patchFeatureFlag_Path;
    headers: patchFeatureFlag_Headers;
    cookie: never;
    body: PatchWithComment;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: FeatureFlag;
          };
        };
        examples: {};
      };
      400: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: InvalidRequestErrorRep;
          };
        };
        examples: {};
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UnauthorizedErrorRep;
          };
        };
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: NotFoundErrorRep;
          };
        };
        examples: {};
      };
      405: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: MethodNotAllowedErrorRep;
          };
        };
        examples: {};
      };
      409: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: StatusConflictErrorRep;
          };
        };
        examples: {};
      };
      429: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: RateLimitedErrorRep;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Delete a feature flag in all environments. Use with caution: only delete feature flags your application no longer uses.
 */
export type deleteFeatureFlag = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: deleteFeatureFlag_Path;
    headers: deleteFeatureFlag_Headers;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UnauthorizedErrorRep;
          };
        };
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: NotFoundErrorRep;
          };
        };
        examples: {};
      };
      409: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: StatusConflictErrorRep;
          };
        };
        examples: {};
      };
      429: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: RateLimitedErrorRep;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type getFeatureFlag_Query = {
  /**
   * Filter configurations by environment
   */
  env?: string;
  /**
   * A comma-separated list of fields to expand in the response. Supported fields are explained above.
   */
  expand?: string;
};

export type getFeatureFlag_Path = {
  /**
   * The project key
   */
  projectKey: string;
  /**
   * The feature flag key
   */
  featureFlagKey: string;
};

export type getFeatureFlag_Headers = { Authorization: string };

export type patchFeatureFlag_Path = {
  /**
   * The project key
   */
  projectKey: string;
  /**
   * The feature flag key. The key identifies the flag in your code.
   */
  featureFlagKey: string;
};

export type patchFeatureFlag_Headers = { Authorization: string };

export type deleteFeatureFlag_Path = {
  /**
   * The project key
   */
  projectKey: string;
  /**
   * The feature flag key. The key identifies the flag in your code.
   */
  featureFlagKey: string;
};

export type deleteFeatureFlag_Headers = { Authorization: string };